+++ saleor/checkout/tests/test_checkout_complete.py
@@ -2,7 +2,6 @@ from decimal import Decimal
 from unittest import mock
 
 import pytest
-from django.contrib.auth.models import AnonymousUser
 from django.test import override_settings
 from prices import TaxedMoney
 
@@ -11,6 +10,7 @@ from ...account.models import CustomerEvent
 from ...core.exceptions import InsufficientStock
 from ...core.notify_events import NotifyEventType
 from ...core.taxes import zero_money, zero_taxed_money
+from ...core.tests.utils import get_site_context_payload
 from ...discount.models import VoucherCustomer
 from ...giftcard import GiftCardEvents
 from ...giftcard.models import GiftCard, GiftCardEvent
@@ -41,6 +41,7 @@ def test_create_order_captured_payment_creates_expected_events(
     shipping_method,
     payment_txn_captured,
     channel_USD,
+    site_settings,
 ):
     checkout = checkout_with_item
     checkout_user = customer_user
@@ -127,8 +128,7 @@ def test_create_order_captured_payment_creates_expected_events(
     expected_order_payload = {
         "order": get_default_order_payload(order, checkout.redirect_url),
         "recipient_email": order.get_customer_email(),
-        "site_name": "mirumee.com",
-        "domain": "mirumee.com",
+        **get_site_context_payload(site_settings.site),
     }
 
     expected_payment_payload = {
@@ -142,8 +142,7 @@ def test_create_order_captured_payment_creates_expected_events(
             "captured_amount": payment_txn_captured.captured_amount,
             "currency": payment_txn_captured.currency,
         },
-        "site_name": "mirumee.com",
-        "domain": "mirumee.com",
+        **get_site_context_payload(site_settings.site),
     }
     # Ensure the correct order confirmed event was created
     # should be order confirmed event
@@ -190,6 +189,7 @@ def test_create_order_captured_payment_creates_expected_events_anonymous_user(
     shipping_method,
     payment_txn_captured,
     channel_USD,
+    site_settings,
 ):
     checkout = checkout_with_item
     checkout_user = None
@@ -223,7 +223,7 @@ def test_create_order_captured_payment_creates_expected_events_anonymous_user(
             discounts=None,
             taxes_included_in_prices=True,
         ),
-        user=AnonymousUser(),
+        user=None,
         app=None,
         manager=manager,
     )
@@ -277,8 +277,7 @@ def test_create_order_captured_payment_creates_expected_events_anonymous_user(
     expected_order_payload = {
         "order": get_default_order_payload(order, checkout.redirect_url),
         "recipient_email": order.get_customer_email(),
-        "site_name": "mirumee.com",
-        "domain": "mirumee.com",
+        **get_site_context_payload(site_settings.site),
     }
 
     expected_payment_payload = {
@@ -292,8 +291,7 @@ def test_create_order_captured_payment_creates_expected_events_anonymous_user(
             "captured_amount": payment_txn_captured.captured_amount,
             "currency": payment_txn_captured.currency,
         },
-        "site_name": "mirumee.com",
-        "domain": "mirumee.com",
+        **get_site_context_payload(site_settings.site),
     }
 
     # Ensure the correct order confirmed event was created
@@ -336,6 +334,7 @@ def test_create_order_preauth_payment_creates_expected_events(
     shipping_method,
     payment_txn_preauth,
     channel_USD,
+    site_settings,
 ):
     checkout = checkout_with_item
     checkout_user = customer_user
@@ -409,8 +408,7 @@ def test_create_order_preauth_payment_creates_expected_events(
     expected_payload = {
         "order": get_default_order_payload(order, checkout.redirect_url),
         "recipient_email": order.get_customer_email(),
-        "site_name": "mirumee.com",
-        "domain": "mirumee.com",
+        **get_site_context_payload(site_settings.site),
     }
 
     # Ensure the correct order confirmed event was created
@@ -448,6 +446,7 @@ def test_create_order_preauth_payment_creates_expected_events_anonymous_user(
     shipping_method,
     payment_txn_preauth,
     channel_USD,
+    site_settings,
 ):
     checkout = checkout_with_item
     checkout_user = None
@@ -481,7 +480,7 @@ def test_create_order_preauth_payment_creates_expected_events_anonymous_user(
             discounts=[],
             taxes_included_in_prices=True,
         ),
-        user=AnonymousUser(),
+        user=None,
         app=None,
         manager=manager,
     )
@@ -522,8 +521,7 @@ def test_create_order_preauth_payment_creates_expected_events_anonymous_user(
     expected_payload = {
         "order": get_default_order_payload(order, checkout.redirect_url),
         "recipient_email": order.get_customer_email(),
-        "site_name": "mirumee.com",
-        "domain": "mirumee.com",
+        **get_site_context_payload(site_settings.site),
     }
     # Ensure the correct order confirmed event was created
     # should be order confirmed event
@@ -661,7 +659,7 @@ def test_create_order_with_gift_card(
             discounts=None,
             taxes_included_in_prices=True,
         ),
-        user=customer_user if not is_anonymous_user else AnonymousUser(),
+        user=customer_user if not is_anonymous_user else None,
         app=None,
         manager=manager,
     )
@@ -867,7 +865,7 @@ def test_create_order_gift_card_bought(
             discounts=None,
             taxes_included_in_prices=True,
         ),
-        user=customer_user if not is_anonymous_user else AnonymousUser(),
+        user=customer_user if not is_anonymous_user else None,
         app=None,
         manager=manager,
     )
@@ -948,7 +946,7 @@ def test_create_order_gift_card_bought_order_not_captured_gift_cards_not_sent(
             discounts=None,
             taxes_included_in_prices=True,
         ),
-        user=customer_user if not is_anonymous_user else AnonymousUser(),
+        user=customer_user if not is_anonymous_user else None,
         app=None,
         manager=manager,
     )
@@ -1012,7 +1010,7 @@ def test_create_order_gift_card_bought_only_shippable_gift_card(
             discounts=None,
             taxes_included_in_prices=True,
         ),
-        user=customer_user if not is_anonymous_user else AnonymousUser(),
+        user=customer_user if not is_anonymous_user else None,
         app=None,
         manager=manager,
     )
@@ -1072,7 +1070,7 @@ def test_create_order_gift_card_bought_do_not_fulfill_gift_cards_automatically(
             discounts=None,
             taxes_included_in_prices=True,
         ),
-        user=customer_user if not is_anonymous_user else AnonymousUser(),
+        user=customer_user if not is_anonymous_user else None,
         app=None,
         manager=manager,
     )
@@ -1191,7 +1189,12 @@ def test_create_order_use_translations(
 
 @mock.patch("saleor.plugins.manager.PluginsManager.notify")
 def test_complete_checkout_0_total_captured_payment_creates_expected_events(
-    mock_notify, checkout_with_item_total_0, customer_user, channel_USD, app
+    mock_notify,
+    checkout_with_item_total_0,
+    customer_user,
+    channel_USD,
+    app,
+    site_settings,
 ):
     checkout = checkout_with_item_total_0
     checkout_user = customer_user
@@ -1255,8 +1258,7 @@ def test_complete_checkout_0_total_captured_payment_creates_expected_events(
     expected_order_payload = {
         "order": get_default_order_payload(order, checkout.redirect_url),
         "recipient_email": order.get_customer_email(),
-        "site_name": "mirumee.com",
-        "domain": "mirumee.com",
+        **get_site_context_payload(site_settings.site),
     }
 
     # Ensure the correct order confirmed event was created