+++ saleor/core/permissions/auth_filters.py
@@ -6,7 +6,8 @@ def is_app(context):
 
 
 def is_user(context):
-    return context.user.is_active and context.user.is_authenticated
+    user = context.user
+    return user and user.is_active
 
 
 def is_staff_user(context):