+++ saleor/graphql/checkout/tests/mutations/test_checkout_complete.py
@@ -5,7 +5,7 @@ from unittest.mock import ANY, patch
 import graphene
 import pytest
 import pytz
-from django.contrib.auth.models import AnonymousUser
+from django.contrib.sites.models import Site
 from django.db.models.aggregates import Sum
 from django.utils import timezone
 from prices import Money
@@ -564,7 +564,7 @@ def test_checkout_complete_by_app_with_missing_permission(
         payment_data=ANY,
         store_source=ANY,
         discounts=ANY,
-        user=AnonymousUser(),
+        user=None,
         app=ANY,
         site_settings=ANY,
         tracking_code=ANY,
@@ -770,6 +770,7 @@ def test_checkout_complete_requires_confirmation(
 ):
     site_settings.automatically_confirm_all_new_orders = False
     site_settings.save()
+    Site.objects.clear_cache()
     payment = payment_dummy
     payment.checkout = checkout_ready_to_complete
     payment.save()