+++ saleor/graphql/core/tests/test_graphql.py
@@ -3,7 +3,6 @@ from unittest.mock import Mock
 
 import graphene
 import pytest
-from django.contrib.auth.models import AnonymousUser
 from django.shortcuts import reverse
 from graphql.error import GraphQLError
 from graphql_relay import to_global_id
@@ -52,14 +51,14 @@ def test_jwt_middleware(client, admin_user):
     response = api_client_post(data={"query": user_details_query})
     repl_data = response.json()
     assert response.status_code == 200
-    assert isinstance(response.wsgi_request.user, AnonymousUser)
+    assert not response.wsgi_request.user
     assert repl_data["data"]["me"] is None
 
     # test creating a token for admin user
     response = api_client_post(data={"query": create_token_query})
     repl_data = response.json()
     assert response.status_code == 200
-    assert isinstance(response.wsgi_request.user, AnonymousUser)
+    assert response.wsgi_request.user == admin_user
     token = repl_data["data"]["tokenCreate"]["token"]
     assert token is not None