+++ saleor/graphql/order/tests/test_order.py
@@ -8,7 +8,6 @@ from unittest.mock import ANY, MagicMock, Mock, call, patch
 import graphene
 import pytest
 import pytz
-from django.contrib.auth.models import AnonymousUser
 from django.core.exceptions import ValidationError
 from django.core.files import File
 from django.db.models import Sum
@@ -25,6 +24,7 @@ from ....core.notify_events import NotifyEventType
 from ....core.postgres import FlatConcatSearchVector
 from ....core.prices import quantize_price
 from ....core.taxes import TaxError, zero_money, zero_taxed_money
+from ....core.tests.utils import get_site_context_payload
 from ....discount.models import OrderDiscount, VoucherChannelListing
 from ....giftcard import GiftCardEvents
 from ....giftcard.events import gift_cards_bought_event, gift_cards_used_in_order_event
@@ -1102,7 +1102,6 @@ def test_order_available_shipping_methods_with_weight_based_shipping_method(
     permission_manage_orders,
     minimum_order_weight_value,
 ):
-
     shipping_method = shipping_method_weight_based
     order = order_line.order
     if minimum_order_weight_value is not None:
@@ -1599,8 +1598,7 @@ def test_order_confirm(
         "recipient_email": order_unconfirmed.user.email,
         "requester_user_id": to_global_id_or_none(staff_api_client.user),
         "requester_app_id": None,
-        "site_name": "mirumee.com",
-        "domain": "mirumee.com",
+        **get_site_context_payload(site_settings.site),
     }
     mocked_notify.assert_called_once_with(
         NotifyEventType.ORDER_CONFIRMED,
@@ -1622,6 +1620,7 @@ def test_order_confirm_without_sku(
     order_unconfirmed,
     permission_manage_orders,
     payment_txn_preauth,
+    site_settings,
 ):
     order_unconfirmed.lines.update(product_sku=None)
     ProductVariant.objects.update(sku=None)
@@ -1661,8 +1660,7 @@ def test_order_confirm_without_sku(
         "recipient_email": order_unconfirmed.user.email,
         "requester_user_id": to_global_id_or_none(staff_api_client.user),
         "requester_app_id": None,
-        "site_name": "mirumee.com",
-        "domain": "mirumee.com",
+        **get_site_context_payload(site_settings.site),
     }
     mocked_notify.assert_called_once_with(
         NotifyEventType.ORDER_CONFIRMED,
@@ -3941,6 +3939,30 @@ def test_draft_order_delete(staff_api_client, permission_manage_orders, draft_or
     )
     with pytest.raises(order._meta.model.DoesNotExist):
         order.refresh_from_db()
+
+
+def test_draft_order_delete_product(
+    app_api_client, permission_manage_products, draft_order
+):
+    query = """
+        mutation DeleteProduct($id: ID!) {
+          productDelete(id: $id) {
+            product {
+              id
+            }
+          }
+        }
+    """
+    order = draft_order
+    line = order.lines.first()
+    product = line.variant.product
+    product_id = graphene.Node.to_global_id("Product", product.id)
+    variables = {"id": product_id}
+    response = app_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_products]
+    )
+    content = get_graphql_content(response)
+    assert content["data"]["productDelete"]["product"]["id"] == product_id
 
 
 @pytest.mark.parametrize(
@@ -6507,7 +6529,7 @@ def test_order_cancel_as_app(
 
     mock_clean_order_cancel.assert_called_once_with(order)
     mock_cancel_order.assert_called_once_with(
-        order=order, user=AnonymousUser(), app=app_api_client.app, manager=ANY
+        order=order, user=None, app=app_api_client.app, manager=ANY
     )
 
 
@@ -6622,8 +6644,7 @@ def test_order_capture(
             "captured_amount": payment.captured_amount,
             "currency": payment.currency,
         },
-        "site_name": "mirumee.com",
-        "domain": "mirumee.com",
+        **get_site_context_payload(site_settings.site),
     }
 
     mocked_notify.assert_called_once_with(
@@ -7189,7 +7210,6 @@ def test_order_refund_with_transaction_action_request_missing_event(
 def test_clean_payment_without_payment_associated_to_order(
     staff_api_client, permission_manage_orders, order, requires_amount, mutation_name
 ):
-
     assert not OrderEvent.objects.exists()
 
     additional_arguments = ", amount: 2" if requires_amount else ""
@@ -8429,7 +8449,7 @@ def test_order_bulk_cancel_as_app(
     assert not data["errors"]
 
     calls = [
-        call(order=order, user=AnonymousUser(), app=app_api_client.app, manager=ANY)
+        call(order=order, user=None, app=app_api_client.app, manager=ANY)
         for order in orders
     ]
 
@@ -9509,7 +9529,6 @@ def test_orders_query_with_filter_by_orders_id(
     permission_manage_orders,
     channel_USD,
 ):
-
     # given
     orders = Order.objects.bulk_create(
         [
@@ -9548,7 +9567,6 @@ def test_orders_query_with_filter_by_old_orders_id(
     permission_manage_orders,
     channel_USD,
 ):
-
     # given
     orders = Order.objects.bulk_create(
         [
@@ -9589,7 +9607,6 @@ def test_orders_query_with_filter_by_old_and_new_orders_id(
     permission_manage_orders,
     channel_USD,
 ):
-
     # given
     orders = Order.objects.bulk_create(
         [
@@ -10526,12 +10543,12 @@ def test_order_resolver_tax_recalculation(
 
     query = (
         """
-    query OrderPrices($id: ID!) {
-        order(id: $id) {
-            %s { net { amount } gross { amount } }
+        query OrderPrices($id: ID!) {
+            order(id: $id) {
+                %s { net { amount } gross { amount } }
+            }
         }
-    }
-    """
+        """
         % price_name
     )
     variables = {"id": order_id}
@@ -10598,14 +10615,14 @@ def test_order_line_resolver_tax_recalculation(
 
     query = (
         """
-    query OrderLinePrices($id: ID!) {
-        order(id: $id) {
-            lines {
-                %s { net { amount } gross { amount } }
+        query OrderLinePrices($id: ID!) {
+            order(id: $id) {
+                lines {
+                    %s { net { amount } gross { amount } }
+                }
             }
         }
-    }
-    """
+        """
         % price_name
     )
     variables = {"id": order_id}
@@ -10636,7 +10653,6 @@ query OrderShippingTaxRate($id: ID!) {
     }
 }
 """
-
 
 ORDER_LINE_TAX_RATE_QUERY = """
 query OrderLineTaxRate($id: ID!) {